Loop works and refactored Arithemtic and Calculate subclasses.

+++ src/commands/Executable.java
@@ -6,5 +6,5 @@ package commands;
  *
  */
 public interface Executable {
-	public int execute();
+	public double execute();
 }
